requirements.txt
Flask
torch
torchvision
open-clip-torch
Pillow
transformers
